pyproject.toml
[project]
name = "exls"
version = "0.3.0"
description = "The exalsius CLI"
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "rich>=13.9.4",
    "typer>=0.15.2",
    "exalsius-api-client @ git+https://github.com/exalsius/exalsius-api-spec.git@v1.21.6#subdirectory=client-sdk",
    "pydantic-settings>=2.10.1",
    "pyyaml>=6.0.2",
    "filelock>=3.17.0",
    "termcolor>=2.4.0",
    "requests>=2.32.3",
    "auth0-python>=4.10.0",
    "keyring>=25.6.0",
    "qrcode[pil]>=8.2",
    "coolname>=2.2.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["exls*"] 

[project.scripts]
exls = "exls.app:app" 

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.pyright]
typeCheckingMode = "strict"

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "pre-commit>=4.1.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.9.9",
    "pyright>=1.1.376",
]
